Feat: [#32] 추가 관리자 로그인 기능
로그인, 로그아웃 세션 추가하였습니다.

+++ module-admin/src/main/java/com/adminapplication/dto/AllReportsResponseDto.java
@@ -7,9 +7,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AllReportsResponseDto {
-    private Integer id; // 신고된 게시글 id
+    private Long id; // 신고된 게시글 id
     private String title;
-    private Integer userId;
+    private Long userId;
     private Status status;
     private Integer qtyOfReport; // 신고 건수
 }
